bzr revid: ovnicraft@gmail.com-20101227052633-hi3op794dkd3pn5m

+++ addons/audittrail/audittrail.py
@@ -106,14 +106,16 @@ class audittrail_rule(osv.osv):
         """
         obj_action = self.pool.get('ir.actions.act_window')
         val_obj = self.pool.get('ir.values')
+        value=''
         #start Loop
         for thisrule in self.browse(cr, uid, ids):
             if thisrule.id in self.__functions:
                 for function in self.__functions[thisrule.id]:
                     setattr(function[0], function[1], function[2])
             w_id = obj_action.search(cr, uid, [('name', '=', 'View Log'), ('res_model', '=', 'audittrail.log'), ('src_model', '=', thisrule.object_id.model)])
-            obj_action.unlink(cr, uid, w_id)
-            value = "ir.actions.act_window" + ',' + str(w_id[0])
+            if w_id:
+                obj_action.unlink(cr, uid, w_id)
+                value = "ir.actions.act_window" + ',' + str(w_id[0])
             val_id = val_obj.search(cr, uid, [('model', '=', thisrule.object_id.model), ('value', '=', value)])
             if val_id:
                 res = ir.ir_del(cr, uid, val_id[0])
@@ -198,6 +200,8 @@ class audittrail_objects_proxy(osv_pool):
         """
         if not context:
             context = {}
+        if field_name in('__last_update','id'):
+            return values            
         pool = pooler.get_pool(cr.dbname)
         field_pool = pool.get('ir.model.fields')
         model_pool = pool.get('ir.model')
@@ -248,6 +252,8 @@ class audittrail_objects_proxy(osv_pool):
         log_line_pool = pool.get('audittrail.log.line')
         #start Loop
         for line in lines:
+            if line['name'] in('__last_update','id'):
+                continue                
             if obj_pool._inherits:
                 inherits_ids = model_pool.search(cr, uid, [('model', '=', obj_pool._inherits.keys()[0])])
                 field_ids = field_pool.search(cr, uid, [('name', '=', line['name']), ('model_id', 'in', (model.id, inherits_ids[0]))])
@@ -279,6 +285,7 @@ class audittrail_objects_proxy(osv_pool):
                     "field_description": field['field_description']
                     }
             line_id = log_line_pool.create(cr, uid, vals)
+            cr.commit()
         #End Loop
         return True
 
@@ -361,7 +368,7 @@ class audittrail_objects_proxy(osv_pool):
                     lines.append(line)
 
                 self.create_log_line(cr, uid, log_id, model, lines)
-
+            cr.commit()
             cr.close()
             return res
